backend/requirements.txt
# FastAPI Backend Requirements
# Intel Unnati - NCERT RAG System

# ============================================
# Web Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# OCR & PDF Processing (Phase 1)
# ============================================
pdf2image==1.16.3
Pillow==10.2.0
pytesseract==0.3.10
PyPDF2==3.0.1

# ============================================
# Text Processing & Cleaning (Phase 2)
# ============================================
regex==2023.12.25
python-Levenshtein==0.25.0

# ============================================
# Embeddings & Vector Store (Phase 4-5)
# ============================================
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.2
transformers==4.36.2

# ============================================
# LLM Integration (Phase 6)
# ============================================
openai==1.7.2
tiktoken==0.5.2

# ============================================
# CPU Optimizations
# ============================================
psutil==5.9.7

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
tqdm==4.66.1

# ============================================
# Logging & Monitoring
# ============================================
loguru==0.7.2

# ============================================
# Development Tools (Optional)
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
